Finishing touches on GUI
