+++ data/settings.ini
@@ -5,8 +5,8 @@ printAllTemplates: yes
 repeat: no
 
 [letterGeneration]
-rare_chance: 8
-double_chance: 13
+rare_chance: 7
+double_chance: 5
 qu_chance: 7
 diagraph_chance: 15
